add Areas.vue, calculate positions in this component and send it through $ref
