+++ src/types.ts
@@ -1,3 +1,5 @@
+import Vue from 'vue'
+
 export enum EDirection {
   // eslint-disable-next-line no-unused-vars
   ArrowLeft = 'left-arrow',
@@ -20,40 +22,13 @@ export enum EKeys {
   Right = 'ArrowRight'
 }
 
-export interface IArrowData {
-  id: number;
-  direction: EDirection
-}
-
-export interface IFlowProps {
+export interface IAreasPositions {
   topGoodArea: number,
   bottomGoodArea: number,
   topExcellentArea: number,
   bottomExcellentArea: number
 }
 
-export class Block {
-  id = Date.now
-  blockClasses = ['left-arrow', 'up-arrow', 'down-arrow', 'right-arrow']
-
-  createItem () {
-    const newItem = document.createElement('div')
-    newItem.setAttribute('class', `${this.className} drop-block`)
-    newItem.setAttribute('id', `${this.id}`)
-    return newItem
-  }
-
-  getRandom = () => {
-    const elements = Math.floor((Math.random() * 20) + 5)
-    const timeout = Math.floor(Math.random() * 2000 + 200)
-    const itemClass = this.blockClasses[Math.floor(Math.random() * this.blockClasses.length)]
-
-    return {
-      elements,
-      timeout,
-      itemClass
-    }
-  }
-
-  className = this.getRandom().itemClass
+export type VAreas = Vue & {
+  calculatePositions: () => IAreasPositions
 }
